cardapion/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==1.0.4
